+++ common/src/main/java/dev/xylonity/companions/client/entity/model/CorneliusModel.java
@@ -2,6 +2,8 @@ package dev.xylonity.companions.client.entity.model;
 
 import dev.xylonity.companions.Companions;
 import dev.xylonity.companions.common.entity.companion.CorneliusEntity;
+import dev.xylonity.companions.common.util.Util;
+import dev.xylonity.companions.config.CompanionsConfig;
 import net.minecraft.resources.ResourceLocation;
 import software.bernie.geckolib.model.GeoModel;
 
@@ -14,7 +16,7 @@ public class CorneliusModel extends GeoModel<CorneliusEntity> {
 
     @Override
     public ResourceLocation getTextureResource(CorneliusEntity animatable) {
-        if (animatable.hasCustomName()) {
+        if (Util.matchesReskinName(animatable, CompanionsConfig.CORNELIUS_RESKIN_NAMES)) {
             return Companions.of("textures/entity/cornelius_reskin.png");
         }