+++ common/src/main/java/dev/xylonity/companions/client/entity/model/ShadeSwordModel.java
@@ -2,6 +2,8 @@ package dev.xylonity.companions.client.entity.model;
 
 import dev.xylonity.companions.Companions;
 import dev.xylonity.companions.common.entity.companion.ShadeSwordEntity;
+import dev.xylonity.companions.common.util.Util;
+import dev.xylonity.companions.config.CompanionsConfig;
 import net.minecraft.resources.ResourceLocation;
 import software.bernie.geckolib.model.GeoModel;
 
@@ -14,7 +16,7 @@ public class ShadeSwordModel extends GeoModel<ShadeSwordEntity> {
 
     @Override
     public ResourceLocation getTextureResource(ShadeSwordEntity animatable) {
-        if (animatable.hasReskin()) {
+        if (Util.matchesReskinName(animatable, CompanionsConfig.SHADOW_SWORD_RESKIN_NAMES)) {
             if (animatable.isBlood()) {
                 return Companions.of("textures/entity/shade_sword_blood_reskin.png");
             }